move helpers into CommandRunner struct
