+++ src/git.rs
@@ -1,2 +1,2 @@
 pub(crate) mod commands;
-pub(crate) mod helpers;
+pub(crate) mod command_runner;
